gtkkeyhash: Fix some Ctrl + Alt accelerators not working on Win32
On Win32, "Ctrl + Alt" (= AltGr) can change the keyval for a key. Because
_gtk_key_hash_lookup compares keyvals rather than keycodes, this meant that
some accelerators involving the Ctrl + Alt modifiers were not detected.
For example, on the Polish Programmers Layout, Ctrl + Alt + C produces ć. This
made it impossible to trigger the accelerator Ctrl + Alt + C on that layout [0].
Since both Ctrl + Alt + C and ć are potentially valid accelerators, we have to
check for and accept both variants. We do this by performing the key translation
twice if Ctrl+Alt is present: Once with Ctrl+Alt modifiers, and once without.
Then we accept either keyval.

[0] https://gitlab.com/inkscape/inkscape/-/issues/3572