icontheme: Search preferred icon variant first
This fixes the problem that regular icons are not inherited when symbolic variant is available in
the selected theme (e.g. in case of Adwaita and AdwaitaLegacy).

Previously, the search logic was limited to search first in all themes only if the requested icon
is symbolic, and then search all variants in the current theme, which could lead that the symbolic
variant is selected from the current theme rather than the regular icon from the inherited theme,
even if the GTK_ICON_LOOKUP_FORCE_REGULAR flag is specified.

The new implementation checks whether symbolic icons are preferred or not based on the first icon
name, and then the search for symbolic and non-symbolic icons is done in separate passes, ensuring
that the most appropriate icon is found.