inscription: Determinine line height better
Take the script of the text into account for determining the
nominal line height we're using. This can make enough of a difference
between different scripts that the text is pushed off to the bottom,
where it is clipped away.

The downside of this is that it is more work every time we measure
the widget. It it turns out to be a problem, we can cache e.g. the
font metrics.

Fixes: #7103