surface: Ignore events for unmapped surface
After unmapping a surface from the client side, there may be other events
queued up on that same surface that the compositor had already emitted before
we could notify it of the unmapping. Handling those events can lead to an
inconsistent state, such as in the following case:

Using a touchpad on mutter (wayland):
* Tap with two fingers to open a context menu
* Quickly release both fingers, then hold one finger down
  - libinput processes this as holding down the right click
* Hover an action that would close the popup
* With the same finger, push down the touchpad to click
  - libinput generates a right-release and left-press in quick succession
  - mutter emits a right-release and left-press in the same frame
  - gtk will handle the sequence
    + GDK_BUTTON_RELEASE right (invokes the action, closes the popup)
    + GDK_BUTTON_PRESS left (takes a grab on the main window)
    + wl_pointer::leave (leaving the popup surface)
    + wl_pointer::enter (entering the main window, resets modifiers)
    + GDK_BUTTON_RELEASE left

The GDK_BUTTON_PRESS event that originated from the popup surface has instead
generated a grab on the main window, before we could process the enter event.
Now the final GDK_BUTTON_RELEASE event finds no prior pressed state, because
the button state is always reset on enter.
The inconsistency causes the implicit grab to never be released.