wayland: set monitor workarea
Not setting workarea leads to certain widgets to fall back to geometry,
which leads to e.g. gtkentrycompletion position its popup lists with
the offset of the adjacent monitors.

The behavior was only noticeable on a monitor which had `x`/`y` position
set in wl_output information resulting in elements being rendered
detached from the main window.

Fixes #2794.