window: Fix moving focus to its parent during unparenting
Previously, gtk_widget_unparent requested an after-paint event where the
focus was moved, and it set the widget's parent to NULL.
As a result, when finding out where to put the focus, it was never put in the
widget's parent widget, because that was always already null.

That caused focus unexpectedly moving say, when you deleted a file in
Nautilus from outside the app, and had it focused.

Nautilus should handle this situation better, but the Gtk logic was wrong as
well, so at least, this part works now.