window: Avoid synthesizing crossing events early when unmapping a Popover
When unmapping a widget with pointerfocus, we want to repick a new focus target
and emit crossing events. However if the focused widget is part of a Popover,
or in general of a GdkNative, that is being unmapped as a whole, then we know
that we will already emit crossing events later, when receiving surface
leave/enter events from the compositor.