gpu: Add shader compilation support for D3D12
Same structure as the Vulkan code:

We keep a hash table of compiled shaders indexed by the values that
differ between shaders.
For some surprising reason, they are almost identical.

The only thing D3D12 doesn't need to differentiate by is the root
signature/pipeline layout, because those would only differ for
static/immutable samplers and those are only necessary for YCbcr
which D3D12 doesn't support.

No users of this API yet, this is just the plumbing.