Win32/IME: Fix crash when ImmGetCompositionStringW returns length <= 0
Some IMEs caused GIMP to crash when the text tool was used. This was
reported on the GIMP bugtracker:

  https://gitlab.gnome.org/GNOME/gimp/-/issues/11518

The reason for the crash was a null-pointer dereference in the internal
function get_pango_attr_list() that occured when ImmGetCompositionStringW()
returned a length of 0. In an earlier version of the code,
get_pango_attr_list() had used alloca() rather than g_malloc(), so the
bug was masked: This is because g_malloc() returns NULL for a
zero-length requested size, while alloca() returns a pointer to a
location on the stack that happens to be readable.

In addition, this commit also fixes a potential crash when negative
values are returned by ImmGetCompositionStringW(). Negative values can
be returned according to MSDN if the IME detects an error or has no data
available. Previously, this would have most likely resulted in an
out-of-memory crash due do the signed value being cast to unsigned in
the argument to g_malloc().