widget: Allow calling gtk_widget_dispose_template more than once
gtk_widget_dispose_template () is intended to be called from the
widget's dispose vfunc, and the dispose vfunc can be run multiple times.
When this happens, the child != NULL assertion would fail.

Instead of removing the assertion, eagerly destroy the hash table the
first time gtk_widget_dispose_template () is called. This way, we keep
the assertion for that first run, and avoid even entering the block and
doing the lookup on subsequent runs.

Signed-off-by: Sergey Bugaev <bugaevc@gmail.com>