a11y atspi: Report masked password instead of plain one
When text is requested via the AtspiText interface's
GetText method, use the display text instead of the
plain text for the case of an editable.

gtk_text_get_display_text also implements start/end offset
handling, so drop the extra logic for those.

For a GtkPasswordEntry, this ensures that the masked
password is reported on the AT-SPI level unless the plain
password is also displayed on screen.

Sample scenario:

1) Run the gtk4-demo "Entry" -> "Password Entry" sample
2) type "secret" into the first password field
3) start Accerciser and select the "Password" object
   in the a11y tree
4) query text using Accerciser's IPython console

Without this commit in place, the plain password
was reported even when only the masked password
was shown on screen:

    In [2]: text_iface = acc.queryText()
    In [3]: text_iface.getText(0, text_iface.characterCount)
    Out[3]: 'secret'

With this commit in place, the masked password
is also reported via AT-SPI:

    In [4]: text_iface = acc.queryText()
    In [5]: text_iface.getText(0, text_iface.characterCount)
    Out[5]: '••••••'

When showing the plain text on screen, it is also
still reported via AT-SPI:

    In [6]: text_iface.getText(0, text_iface.characterCount)
    Out[6]: 'secret'

Not reporting the plain text password when not shown on
screen also matches the behavior in GTK 3, as can be seen
e.g. in the gtk3-demo "Entry" -> "Entry Buffer" example.