gtk4-demo: Added GNotifications demo.
This demo showcases the capabilities of the GNotifications object, by
showing a "messaging app simulation".

The user can receive any kind of notifications, with different
priorities, such as urgent priority (incoming calls), normal priority
(incoming messages), and it also has a silent mode, which enables
notifications with low priorities for both calls and messages.

This demo also allows user to click notifications with the app closed,
which will open the app if it's closed, this is achieved thanks to DBus
activatabilty.

New meta-files:

  - gtk-demo/org.gtk.Demo4.App.desktop This file is required in order
  for notifications to be sended.

  - gtk-demo/org.gtk.Demo4.App.service.in This file is required in order
  for DBus activatabilty to work.

Some changes to the application.c demo:

  - As Matthias Clasen suggested, this demo got merged into
  GtkApplication demo executable, so that no new executable would be
  required, but it's in a way so that the demo looks like it's
  standalone.

Closes (#7610)