application: Rework host app portal registration
The previous implementation had a hole where host apps could sometimes
remain unregistered. Our Register() call was configured to not
socket-activate the portal. If the app launches early enough, the portal
isn't yet running and so the Register() call just disappears

The previous implementation also ignored the various ways to disable
portals, like $GDK_DEBUG or a call to gtk_disable_portals

Essentially, to do this right we would need to call
gdk_display_should_use_portal to see if we should call Register(), and
if we do allow it to socket-activate the portal. However, we cannot do
that because we haven't yet called gtk_init

This implementation hinges on the fact that we always call
gdk_display_should_use_portal before actually using any portal. This
means that we can place the Register() call there and solve both issues
at once.