Numpad keys behave differently than non numpad keys when combined with shift
Author: Brian Kress

With numlock off, on non GTK apps holding shift then using a numpad key will extend the selection as the cursor moves.  With GTK apps this numpad keys will not extend the selection with shift held, only non numpad keys will.  This problem only occurs on Wayland, the combination is handled correctly on X11.

An investigation into this problem was done in bug #6660 and it was determined this is related to the wayland backend setting consume modifiers in translated events incorrectly.  The current code is using the (deprecated) function xkb_state_mod_mask_remove_consumed.  This function only supports XKB_CONSUMED_MODE_XKB, which seems to be incorrect for GTK.  Changing this code to use xkb_state_key_get_consumed_mods2 allows us to specify XKB_CONSUMED_MODE_GTK instead, which seems to be more correct.

With this change shift and numpad keys is able to extend the selection, matching X11 GTK apps and non GTK apps.

Fixes #6660

This is backport of <!8239> to gtk 3.24