gtk: resize/redraw a GtkToolItemGroup when any of its children changes visibility.
In particular, GtkToolItem widgets have 2 specific properties for
visibility depending on the orientation, and this is used in computing
the group's size, and obviously in drawing it. But a change of these
properties on a children was not triggering anything, so unless some
other change would trigger a redraw as a side effect, such change was
not taken into account.