gsk: Support repeating conic gradients
The GskGradient struct has repeat modes already.
This commit adds a way to map a subset of the full circle
to the [0,1] range of the gradient, and handles the rest
of the circle according to the repeat mode.

The new properties for conic-gradient in node files are
repeat, start and end.