imcontextwayland: Translate cursor rectangle to correct native surface
Currently, when preediting within an editable area in a popover
surface, the input panel was shifted to the lower right of the caret
due to erroneous coordinate translation. In notify_cursor_location(),
the position passed to zwp_text_input_v3_set_cursor_rectangle() ended
up relative to GtkRoot instead of the actual GtkNative containing the
editable area, which led to an extra offset that amounts to the
difference between the popover coordinates and the toplevel coordinates.

This patch translates the cursor rectangle to the native surface of the
widget.

Closes #7833